The default behavior in Vulkan is to use blend_pass_through, which "Each fragment shader output will have its value directly written to the framebuffer attachment. This is the default."
Default behavior leaves jaggies although the font texture appropriately has none.

Sub-pixel support in rusttype is no known or addressed in this commit, but fiddling with the provided shader showed only the R channel to be non-zero.

The documents for vulkano also include `blend_collective` and this was also effective at achieving proper alpha blending into the background.  If you are reading git blame and suspect that the alpha compositing is not meeting your needs, investigate the other pipeline blending methods for vulkano::pipeline::PipelineBuilder